% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oct_2024_all_nonfarm}
\alias{Oct_2024_all_nonfarm}
\title{Oct_2024_all_nonfarm}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 118 rows and 2 columns.
}
\usage{
Oct_2024_all_nonfarm
}
\description{
This is a report of all nonfarm employees in the United States, as reported by the Bureau of Labor Statistics. The report runs from January, 2015 through October, 2024
}
\details{
\describe{
 \item{Period}{The month for the subject of the labor survey}
 \item{Value}{The number of people working}
 }

 @source https://data.bls.gov/dataViewer/view/timeseries/CES0000000001
}
\keyword{datasets}
