% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oct_2024_avg_hourly_pay}
\alias{Oct_2024_avg_hourly_pay}
\title{Oct_2024_avg_hourly_pay}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 117 rows and 2 columns.
}
\usage{
Oct_2024_avg_hourly_pay
}
\description{
This is a report of the average hourly pay for all workers in the United States, as reported by the Bureau of Labor Statistics
}
\details{
\describe{
 \item{Period}{The month for the subject of the labor survey}
 \item{Value}{The average hourly pay in the United States}
 }

 @source https://data.bls.gov/dataViewer/view/timeseries/CES0500000003
}
\keyword{datasets}
