% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oct_2024_unemployment}
\alias{Oct_2024_unemployment}
\title{Oct_2024_unemployment}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 117 rows and 2 columns.
}
\usage{
Oct_2024_unemployment
}
\description{
This is a report of the unemployment rate in the United States, as reported by the Bureau of Labor Statistics
}
\details{
\describe{
 \item{Period}{The month for the subject of the labor survey}
 \item{Value}{The unemployment rate}
 }
 @source https://data.bls.gov/dataViewer/view/timeseries/LNS14000000
}
\keyword{datasets}
